day 10 part 2
folded and used a solver library for this. The one I chose actually has
a bug where integer variables aren't written to the LP file that HiGHs
sees so I have included a patch in this repo to apply to the dependency
if I want to run this again in future. (deps are exluded in gitignore
and would lose the manual changes on rebuild)